Git is Hub

+++ zsh/aliases.zsh
@@ -23,12 +23,14 @@ ialias rm='trash-put' # Requires trash-cli
 ialias dd='dcfldd' # DD with a progress report
 
 # Apt stuff
-ialias agi='sudo apt-get install'
-ialias agp='sudo apt-get purge'
-ialias agr='sudo apt-get remove'
-ialias agu='sudo apt-get update'
-ialias agug='sudo apt-get upgrade'
-ialias aguu='sudo apt-get update && sudo apt-get upgrade'
+if which apt-get &> /dev/null; then
+  alias agi='sudo apt-get install'
+  alias agp='sudo apt-get purge'
+  alias agr='sudo apt-get remove'
+  alias agu='sudo apt-get update'
+  alias agug='sudo apt-get upgrade'
+  alias aguu='sudo apt-get update && sudo apt-get upgrade'
+fi
 
 alias xtitle='set_terminal_title'
 
@@ -38,3 +40,5 @@ alias ping='ping -c 5'
 alias clr='clear;echo "Currently logged in on $(tty), as $USER in directory $PWD."'
 alias path='echo -e ${PATH//:/\\n}'
 alias mkdir='mkdir -pv'
+
+ialias git='hub'
